Remove and ignore Gemfile.lock
Them gem itself doesn't need one, and test applications need to be
bundled before tests can be run anyways.

Keeping a Gemfile.lock in the application directory means that the file
will be updated whenever you run tests, and receive the version of the
previous release if you're doing a new one. No point in that. :)